context: components segmented and reorganized ready for state interaction
