+++ src/stores/AppReducer.tsx
@@ -1,0 +1,6 @@
+export default (state: any, action: any) => {
+    switch(action.type) {
+        default:
+            return state;
+    }
+}
